:new: Move exec function to helpers.js
