+++ src/types.js
@@ -11,3 +11,14 @@ export type TempFiles = {
   name: string,
   contents: string
 }
+
+export type ExecResult = string | {
+  stdout: string,
+  stderr: string
+}
+
+export type ExecOptions = {
+  env?: Object,
+  stream?: 'stdout' | 'stderr' | 'both',
+  throwOnStdErr?: boolean
+}
